Remove abstraxion library, upgrade cosmjs libs (#85)
* Remove ""@burnt-labs/abstraxion" dep

Signed-off-by: Justin <328965+justinbarry@users.noreply.github.com>

* Remove Landing component

Deleted the Landing component as it was no longer needed. This cleanup helps reduce the codebase and improves maintainability.

Signed-off-by: Justin <328965+justinbarry@users.noreply.github.com>

* Update cosmjs dependencies to version 0.32.4

Updated @cosmjs/cosmwasm-stargate, @cosmjs/crypto, @cosmjs/proto-signing, and @cosmjs/stargate to version 0.32.4 in both package.json and package-lock.json. This ensures the project benefits from the latest bug fixes and performance improvements offered in the updated libraries.

Signed-off-by: Justin <328965+justinbarry@users.noreply.github.com>

---------

Signed-off-by: Justin <328965+justinbarry@users.noreply.github.com>